ia-models/requirements_cdr.txt
rembg[gpu] # Versión optimizada (quita [gpu] si no tienes NVIDIA)
pillow
numpy
onnxruntime # Requerido por rembg para inferencia
